twitter-scraper/requirements.txt
patchright
python-dotenv>=1.0.0
supabase>=2.0.0
